# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/api/jwt/check": {
            "get": {
                "security": [
                    {
                        "jwt-access": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jwt-check"
                ],
                "responses": {
                    "200": {
                        "description": "comment",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-login"
                ],
                "parameters": [
                    {
                        "description": "User Info Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/common.LoginInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.LoginResult"
                        }
                    }
                }
            }
        },
        "/api/refresh": {
            "get": {
                "security": [
                    {
                        "jwt-refresh": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jwt-refresh"
                ],
                "responses": {
                    "200": {
                        "description": "comment",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.LoginInfo": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "common.LoginResult": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_agree": {
                    "type": "boolean"
                },
                "jwt": {
                    "$ref": "#/definitions/jwt.TokenDetails"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "sms_agree": {
                    "type": "boolean"
                },
                "status": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "jwt.TokenDetails": {
            "type": "object",
            "properties": {
                "accesstoken": {
                    "type": "string"
                },
                "accessuuid": {
                    "type": "string"
                },
                "atexpires": {
                    "type": "integer"
                },
                "refreshtoken": {
                    "type": "string"
                },
                "refreshuuid": {
                    "type": "string"
                },
                "rtexpires": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "jwt-access": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "jwt-refresh": {
            "type": "apiKey",
            "name": "Refresh_Token",
            "in": "header"
        }
    }
}